+++ src/test/java/com/spotride/spotride/model/user/UserControllerTest.java
@@ -1,12 +1,10 @@
-package com.spotride.spotride.user.service;
+package com.spotride.spotride.model.user;
 
-import com.spotride.spotride.model.user.UserMapper;
-import com.spotride.spotride.model.user.service.UserService;
-import com.spotride.spotride.model.user.UserMapperImpl;
+import com.spotride.spotride.model.user.controller.UserController;
+import com.spotride.spotride.model.user.dto.UserResponseDto;
 import com.spotride.spotride.model.user.dto.request.UserCreateRequestDto;
 import com.spotride.spotride.model.user.dto.request.UserUpdateRequestDto;
-import com.spotride.spotride.model.user.model.User;
-import com.spotride.spotride.model.user.repository.UserRepository;
+import com.spotride.spotride.model.user.service.UserService;
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -16,85 +14,73 @@ import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 
 import java.time.LocalDate;
-import java.time.LocalDateTime;
 import java.util.List;
-import java.util.Optional;
+import java.util.Objects;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
-import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.doNothing;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 /**
- * Tests for {@link UserService}.
+ * Tests for {@link UserController}.
  */
-@ContextConfiguration(classes = {UserService.class, UserMapperImpl.class})
+@ContextConfiguration(classes = UserController.class)
 @ExtendWith(SpringExtension.class)
 @SuppressFBWarnings(value = "SECHCP", justification = "Hardcoded password.")
-class UserServiceTest {
-
-    private static final LocalDateTime DATE_TIME_NOW = LocalDateTime.now();
+class UserControllerTest {
 
     @MockBean
-    private UserRepository mockUserRepository;
+    private UserService mockUserService;
 
     @Autowired
-    private UserMapper userMapper;
-
-    @Autowired
-    private UserService userService;
+    private UserController userController;
 
     @Test
     void testGetAllUsers() {
-        var user = User.builder()
+        var userDto = UserResponseDto.builder()
                 .id(1L)
                 .username("john")
-                .password("password")
                 .email("john@example.com")
                 .firstName("John")
                 .lastName("Doe")
                 .phoneNumber("123456789")
                 .birthDate(LocalDate.now())
                 .city("CityName")
-                .createdAt(DATE_TIME_NOW)
-                .modifiedAt(null)
                 .build();
 
-        when(mockUserRepository.findAll()).thenReturn(List.of(user));
+        when(mockUserService.getAll()).thenReturn(List.of(userDto));
 
-        var result = userService.getAll();
+        var result = userController.getAll();
 
+        assertNotNull(result);
         assertEquals(1, result.size());
-        assertEquals("john", result.getFirst().getUsername());
-        verify(mockUserRepository, times(1)).findAll();
+        verify(mockUserService, times(1)).getAll();
     }
 
     @Test
     void testGetUserById() {
-        var user = User.builder()
+        var userDto = UserResponseDto.builder()
                 .id(1L)
                 .username("john")
-                .password("password")
                 .email("john@example.com")
                 .firstName("John")
                 .lastName("Doe")
                 .phoneNumber("123456789")
                 .birthDate(LocalDate.now())
                 .city("CityName")
-                .createdAt(DATE_TIME_NOW)
-                .modifiedAt(null)
                 .build();
 
-        when(mockUserRepository.findById(1L)).thenReturn(Optional.of(user));
+        when(mockUserService.getById(1L)).thenReturn(userDto);
 
-        var result = userService.getById(1L);
+        var response = userController.getById(1L);
 
-        assertNotNull(result);
-        assertEquals("john", result.getUsername());
-        verify(mockUserRepository, times(1)).findById(1L);
+        assertNotNull(response);
+        assertEquals(200, response.getStatusCode().value());
+        assertEquals("john", Objects.requireNonNull(response.getBody()).getUsername());
+        verify(mockUserService, times(1)).getById(1L);
     }
 
     @Test
@@ -110,29 +96,26 @@ class UserServiceTest {
                 .city("CityName")
                 .build();
 
-        var savedUser = User.builder()
+        var createdUserDto = UserResponseDto.builder()
                 .id(1L)
                 .username("john")
-                .password("password")
                 .email("john@example.com")
                 .firstName("John")
                 .lastName("Doe")
                 .phoneNumber("123456789")
                 .birthDate(LocalDate.now())
                 .city("CityName")
-                .createdAt(DATE_TIME_NOW)
-                .modifiedAt(DATE_TIME_NOW)
                 .build();
 
-        when(mockUserRepository.save(any(User.class))).thenReturn(savedUser);
 
-        var user = userMapper.toEntity(userCreateRequestDto);
-        var createdUser = userService.create(userCreateRequestDto);
+        when(mockUserService.create(userCreateRequestDto)).thenReturn(createdUserDto);
 
-        assertNotNull(createdUser);
-        assertEquals("john", createdUser.getUsername());
-        assertEquals("john@example.com", createdUser.getEmail());
-        verify(mockUserRepository, times(1)).save(user);
+        var response = userController.create(userCreateRequestDto);
+
+        assertNotNull(response);
+        assertEquals(200, response.getStatusCode().value());
+        assertEquals("john", Objects.requireNonNull(response.getBody()).getUsername());
+        verify(mockUserService, times(1)).create(userCreateRequestDto);
     }
 
     @Test
@@ -149,51 +132,35 @@ class UserServiceTest {
                 .city("CityName")
                 .build();
 
-        var user = User.builder()
-                .id(1L)
-                .username("john")
-                .password("password")
-                .email("john@example.com")
-                .firstName("John")
-                .lastName("Doe")
-                .phoneNumber("123456789")
-                .birthDate(LocalDate.now())
-                .city("CityName")
-                .createdAt(DATE_TIME_NOW)
-                .modifiedAt(null)
-                .build();
-
-        var updatedUser = User.builder()
+        var updatedUserDto = UserResponseDto.builder()
                 .id(1L)
                 .username("john_updated")
-                .password("password")
                 .email("john_updated@example.com")
                 .firstName("John")
                 .lastName("Doe")
                 .phoneNumber("123456789")
                 .birthDate(LocalDate.now())
                 .city("CityName")
-                .createdAt(DATE_TIME_NOW)
-                .modifiedAt(DATE_TIME_NOW)
                 .build();
 
-        when(mockUserRepository.findById(1L)).thenReturn(Optional.of(user));
-        when(mockUserRepository.save(user)).thenReturn(updatedUser);
+        when(mockUserService.update(1L, userUpdateRequestDto)).thenReturn(updatedUserDto);
 
-        userMapper.updateEntityFromDto(userUpdateRequestDto, user);
-        var result = userService.update(1L, userUpdateRequestDto);
+        var response = userController.update(1L, userUpdateRequestDto);
 
-        assertNotNull(result);
-        assertEquals("john_updated", result.getUsername());
-        verify(mockUserRepository, times(1)).save(user);
+        assertNotNull(response);
+        assertEquals(200, response.getStatusCode().value());
+        assertEquals("john_updated", Objects.requireNonNull(response.getBody()).getUsername());
+        verify(mockUserService, times(1)).update(1L, userUpdateRequestDto);
     }
 
     @Test
     void testDeleteUser() {
-        doNothing().when(mockUserRepository).deleteById(1L);
+        doNothing().when(mockUserService).delete(1L);
 
-        userService.delete(1L);
+        var response = userController.delete(1L);
 
-        verify(mockUserRepository, times(1)).deleteById(1L);
+        assertNotNull(response);
+        assertEquals(204, response.getStatusCode().value());
+        verify(mockUserService, times(1)).delete(1L);
     }
 }
