+++ src/test/java/com/spotride/spotride/model/user/UserServiceTest.java
@@ -1,10 +1,10 @@
-package com.spotride.spotride.user.controller;
+package com.spotride.spotride.model.user;
 
-import com.spotride.spotride.model.user.controller.UserController;
-import com.spotride.spotride.model.user.dto.UserResponseDto;
+import com.spotride.spotride.model.user.service.UserService;
 import com.spotride.spotride.model.user.dto.request.UserCreateRequestDto;
 import com.spotride.spotride.model.user.dto.request.UserUpdateRequestDto;
-import com.spotride.spotride.model.user.service.UserService;
+import com.spotride.spotride.model.user.model.User;
+import com.spotride.spotride.model.user.repository.UserRepository;
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -14,73 +14,85 @@ import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 
 import java.time.LocalDate;
+import java.time.LocalDateTime;
 import java.util.List;
-import java.util.Objects;
+import java.util.Optional;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.doNothing;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 /**
- * Tests for {@link UserController}.
+ * Tests for {@link UserService}.
  */
-@ContextConfiguration(classes = UserController.class)
+@ContextConfiguration(classes = {UserService.class, UserMapperImpl.class})
 @ExtendWith(SpringExtension.class)
 @SuppressFBWarnings(value = "SECHCP", justification = "Hardcoded password.")
-class UserControllerTest {
+class UserServiceTest {
+
+    private static final LocalDateTime DATE_TIME_NOW = LocalDateTime.now();
 
     @MockBean
-    private UserService mockUserService;
+    private UserRepository mockUserRepository;
 
     @Autowired
-    private UserController userController;
+    private UserMapper userMapper;
+
+    @Autowired
+    private UserService userService;
 
     @Test
     void testGetAllUsers() {
-        var userDto = UserResponseDto.builder()
+        var user = User.builder()
                 .id(1L)
                 .username("john")
+                .password("password")
                 .email("john@example.com")
                 .firstName("John")
                 .lastName("Doe")
                 .phoneNumber("123456789")
                 .birthDate(LocalDate.now())
                 .city("CityName")
+                .createdAt(DATE_TIME_NOW)
+                .modifiedAt(null)
                 .build();
 
-        when(mockUserService.getAll()).thenReturn(List.of(userDto));
+        when(mockUserRepository.findAll()).thenReturn(List.of(user));
 
-        var result = userController.getAll();
+        var result = userService.getAll();
 
-        assertNotNull(result);
         assertEquals(1, result.size());
-        verify(mockUserService, times(1)).getAll();
+        assertEquals("john", result.getFirst().getUsername());
+        verify(mockUserRepository, times(1)).findAll();
     }
 
     @Test
     void testGetUserById() {
-        var userDto = UserResponseDto.builder()
+        var user = User.builder()
                 .id(1L)
                 .username("john")
+                .password("password")
                 .email("john@example.com")
                 .firstName("John")
                 .lastName("Doe")
                 .phoneNumber("123456789")
                 .birthDate(LocalDate.now())
                 .city("CityName")
+                .createdAt(DATE_TIME_NOW)
+                .modifiedAt(null)
                 .build();
 
-        when(mockUserService.getById(1L)).thenReturn(userDto);
+        when(mockUserRepository.findById(1L)).thenReturn(Optional.of(user));
 
-        var response = userController.getById(1L);
+        var result = userService.getById(1L);
 
-        assertNotNull(response);
-        assertEquals(200, response.getStatusCode().value());
-        assertEquals("john", Objects.requireNonNull(response.getBody()).getUsername());
-        verify(mockUserService, times(1)).getById(1L);
+        assertNotNull(result);
+        assertEquals("john", result.getUsername());
+        verify(mockUserRepository, times(1)).findById(1L);
     }
 
     @Test
@@ -96,26 +108,29 @@ class UserControllerTest {
                 .city("CityName")
                 .build();
 
-        var createdUserDto = UserResponseDto.builder()
+        var savedUser = User.builder()
                 .id(1L)
                 .username("john")
+                .password("password")
                 .email("john@example.com")
                 .firstName("John")
                 .lastName("Doe")
                 .phoneNumber("123456789")
                 .birthDate(LocalDate.now())
                 .city("CityName")
+                .createdAt(DATE_TIME_NOW)
+                .modifiedAt(DATE_TIME_NOW)
                 .build();
 
+        when(mockUserRepository.save(any(User.class))).thenReturn(savedUser);
 
-        when(mockUserService.create(userCreateRequestDto)).thenReturn(createdUserDto);
+        var user = userMapper.toEntity(userCreateRequestDto);
+        var createdUser = userService.create(userCreateRequestDto);
 
-        var response = userController.create(userCreateRequestDto);
-
-        assertNotNull(response);
-        assertEquals(200, response.getStatusCode().value());
-        assertEquals("john", Objects.requireNonNull(response.getBody()).getUsername());
-        verify(mockUserService, times(1)).create(userCreateRequestDto);
+        assertNotNull(createdUser);
+        assertEquals("john", createdUser.getUsername());
+        assertEquals("john@example.com", createdUser.getEmail());
+        verify(mockUserRepository, times(1)).save(user);
     }
 
     @Test
@@ -132,35 +147,51 @@ class UserControllerTest {
                 .city("CityName")
                 .build();
 
-        var updatedUserDto = UserResponseDto.builder()
+        var user = User.builder()
+                .id(1L)
+                .username("john")
+                .password("password")
+                .email("john@example.com")
+                .firstName("John")
+                .lastName("Doe")
+                .phoneNumber("123456789")
+                .birthDate(LocalDate.now())
+                .city("CityName")
+                .createdAt(DATE_TIME_NOW)
+                .modifiedAt(null)
+                .build();
+
+        var updatedUser = User.builder()
                 .id(1L)
                 .username("john_updated")
+                .password("password")
                 .email("john_updated@example.com")
                 .firstName("John")
                 .lastName("Doe")
                 .phoneNumber("123456789")
                 .birthDate(LocalDate.now())
                 .city("CityName")
+                .createdAt(DATE_TIME_NOW)
+                .modifiedAt(DATE_TIME_NOW)
                 .build();
 
-        when(mockUserService.update(1L, userUpdateRequestDto)).thenReturn(updatedUserDto);
+        when(mockUserRepository.findById(1L)).thenReturn(Optional.of(user));
+        when(mockUserRepository.save(user)).thenReturn(updatedUser);
 
-        var response = userController.update(1L, userUpdateRequestDto);
+        userMapper.updateEntityFromDto(userUpdateRequestDto, user);
+        var result = userService.update(1L, userUpdateRequestDto);
 
-        assertNotNull(response);
-        assertEquals(200, response.getStatusCode().value());
-        assertEquals("john_updated", Objects.requireNonNull(response.getBody()).getUsername());
-        verify(mockUserService, times(1)).update(1L, userUpdateRequestDto);
+        assertNotNull(result);
+        assertEquals("john_updated", result.getUsername());
+        verify(mockUserRepository, times(1)).save(user);
     }
 
     @Test
     void testDeleteUser() {
-        doNothing().when(mockUserService).delete(1L);
+        doNothing().when(mockUserRepository).deleteById(1L);
 
-        var response = userController.delete(1L);
+        userService.delete(1L);
 
-        assertNotNull(response);
-        assertEquals(204, response.getStatusCode().value());
-        verify(mockUserService, times(1)).delete(1L);
+        verify(mockUserRepository, times(1)).deleteById(1L);
     }
 }
